adicionando not found exceptions

+++ API/Equoterapia/src/main/java/com/equoterapia/web/services/SessionService.java
@@ -1,6 +1,7 @@
 package com.equoterapia.web.services;
 
 import com.equoterapia.web.entities.Session;
+import com.equoterapia.web.exceptions.NotFoundException;
 import com.equoterapia.web.repositories.SessionRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -19,25 +20,31 @@ public class SessionService {
     }
     
     public Session findById(Long id) {
+        if (!sessionRepository.existsById(id)){
+            throw new NotFoundException("Sessao não encontrada");
+        }
         Optional<Session> optionalSession = sessionRepository.findById(id);
-        return optionalSession.orElseThrow(() -> new RuntimeException("Sessão não encontrada"));
+        return optionalSession.orElseThrow(() -> new RuntimeException("Runtime Error"));
     }
     
     public Session insert(Session session) {
         return sessionRepository.save(session);
     }
 
- // para verificar se a sessão existe antes de atualizar
+    // para verificar se a sessão existe antes de atualizar
     public Session update(Session session) {
-       
-        findById(session.getId());
+        if (!sessionRepository.existsById(session.getId())){
+            throw new NotFoundException("Sessao não encontrada");
+        }
+
         return sessionRepository.save(session);
     }
 
-//verifica se a sessão existe antes de deletar
+    //verifica se a sessão existe antes de deletar
     public void delete(Long id) {
-        
-        findById(id);
+        if (!sessionRepository.existsById(id)){
+            throw new NotFoundException("Sessao não encontrada");
+        }
         sessionRepository.deleteById(id);
     }
 }
